Incorporate feedback

+++ enterprise/cmd/repo-updater/shared/shared.go
@@ -66,7 +66,7 @@ func EnterpriseInit(
 		// If the feature flag is enabled, create job...
 		if permssync.PermissionSyncWorkerEnabled(ctx, db, observationCtx.Logger) {
 			opts := ossDB.PermissionSyncJobOpts{HighPriority: true}
-			return permsJobStore.CreateRepoSyncJob(ctx, int32(repo), opts)
+			return permsJobStore.CreateRepoSyncJob(ctx, repo, opts)
 		}
 		// .. otherwise we just call the PermsSyncer
 		permsSyncer.ScheduleRepos(ctx, repo)
